+++ pkg/deployments/tasks/deprecated/20220404-erc4626-linear-pool-v2/input.ts
@@ -1,6 +1,6 @@
-import Task, { TaskMode } from '../../src/task';
+import Task, { TaskMode } from '../../../src/task';
 
-export type StablePoolV2Deployment = {
+export type ERC4626LinearPoolDeployment = {
   Vault: string;
 };