+++ pkg/governance-scripts/contracts/TimelockAuthorizerTransitionMigrator.sol
@@ -30,6 +30,7 @@ contract TimelockAuthorizerTransitionMigrator {
     IBasicAuthorizer public immutable oldAuthorizer;
     TimelockAuthorizer public immutable timelockAuthorizer;
     RoleData[] public rolesData;
+    uint256[] public scheduledExecutionIds;
 
     struct RoleData {
         address grantee;
@@ -75,6 +76,7 @@ contract TimelockAuthorizerTransitionMigrator {
         _migrationCompleted = true;
 
         uint256 rolesDataLength = rolesData.length;
+        address[] memory executors = _arr(address(this));
 
         for (uint256 i = 0; i < rolesDataLength; ++i) {
             RoleData memory roleData = rolesData[i];
@@ -83,7 +85,19 @@ contract TimelockAuthorizerTransitionMigrator {
             // The timelock authorizer will emit an event for each permission granted, so we just log the ones we are
             // skipping (if any).
             if (oldAuthorizer.canPerform(roleData.role, roleData.grantee, roleData.target)) {
-                timelockAuthorizer.grantPermissions(_arr(roleData.role), roleData.grantee, _arr(roleData.target));
+                // Check whether the current action has a delay
+                bytes32 roleGrantPermissionActionId = timelockAuthorizer.getGrantPermissionActionId(roleData.role);
+                if (timelockAuthorizer.getActionIdDelay(roleGrantPermissionActionId) == 0) {
+                    timelockAuthorizer.grantPermissions(_arr(roleData.role), roleData.grantee, _arr(roleData.target));
+                } else {
+                    uint256 permissionId = timelockAuthorizer.scheduleGrantPermission(
+                        roleData.role,
+                        roleData.grantee,
+                        roleData.target,
+                        executors
+                    );
+                    scheduledExecutionIds.push(permissionId);
+                }
             } else {
                 emit PermissionSkipped(roleData.role, roleData.grantee, roleData.target);
             }
@@ -95,6 +109,32 @@ contract TimelockAuthorizerTransitionMigrator {
             timelockAuthorizer.EVERYWHERE(),
             false
         );
+    }
+
+    /**
+     * @notice Executes permissions scheduled during migration, all at once.
+     * @dev `migratePermissions` must be called successfully first.
+     * Execution IDs that cannot be executed at this time for any reason (delay not yet due, action was canceled, or
+     * action already executed) are skipped.
+     *
+     * This function can be called more than once without reverting. The execution IDs may have different delays
+     * associated with them, and this function can execute a subset of the execution IDs if it is called before the
+     * longest delay is due.
+     *
+     * On the other hand, this function can still be called after all scheduled execution IDs are resolved (either
+     * cancelled or executed). In that case the function call will do nothing, since the `TimelockAuthorizer` will not
+     * allow any execution.
+     */
+    function executeDelays() external {
+        require(_migrationCompleted, "MIGRATION_INCOMPLETE");
+
+        uint256 scheduledExecutionIdsLength = scheduledExecutionIds.length;
+        for (uint256 i = 0; i < scheduledExecutionIdsLength; ++i) {
+            uint256 scheduledExecutionId = scheduledExecutionIds[i];
+            if (timelockAuthorizer.canExecute(scheduledExecutionId)) {
+                timelockAuthorizer.execute(scheduledExecutionId);
+            }
+        }
     }
 
     // Helper functions