+++ pkg/deployments/CHANGELOG.md
@@ -16,6 +16,12 @@
 - Deployed `AaveLinearPoolFactory` to Gnosis.
 - Deployed `GaugeAdderV3` to Mainnet and Goerli.
 
+### Deprecations
+
+- Deprecated `20221207-aave-rebalanced-linear-pool-v3`.
+- Deprecated `20220425-unbutton-aave-linear-pool`.
+- Deprecated `20220404-erc4626-linear-pool-v2`.
+
 ### Breaking Changes
 
 - The `20221115-aave-rebalanced-linear-pool` task was deleted and replaced with `20221207-aave-rebalanced-linear-pool-v3`. The old task had not been used by anyone.